apply Amines suggestions
