+++ .env.test
@@ -18,7 +18,7 @@ FRANCE_TRAVAIL_AUTH_URL=https://somefakeauthurl.fr
 FRANCE_TRAVAIL_API_URL=https://francetravailfakerdvurl.fr
 FRANCE_TRAVAIL_CLIENT_ID=client_id
 FRANCE_TRAVAIL_CLIENT_SECRET=client_secret
-FRANCE_TRAVAIL_CD_NUMBER=83
+FRANCE_TRAVAIL_WEBHOOKS_DEPARTMENTS=83
 
 AGENT_SIGNATURE_KEY=bc995863-5c80-43a3-a31d-0da216e814a4
 HOST=http://www.rdv-insertion-test.fake
