+++ app/jobs/outgoing_webhooks/france_travail/base_job.rb
@@ -1,0 +1,17 @@
+module OutgoingWebhooks
+  module FranceTravail
+    class BaseJob < ApplicationJob
+      include LockedAndOrderedJobs
+
+      discard_on FranceTravailApi::RetrieveUserToken::UserNotFound
+
+      def self.lock_key(participation_id:, **)
+        "#{base_lock_key}:#{participation_id}"
+      end
+
+      def self.job_timestamp(timestamp:, **)
+        timestamp
+      end
+    end
+  end
+end
