+++ app/jobs/outgoing_webhooks/france_travail/create_participation_job.rb
@@ -1,12 +1,6 @@
 module OutgoingWebhooks
   module FranceTravail
-    class CreateParticipationJob < LockedAndOrderedJobBase
-      discard_on FranceTravailApi::RetrieveUserToken::UserNotFound
-
-      def self.lock_key(participation_id:, **)
-        "#{base_lock_key}:#{participation_id}"
-      end
-
+    class CreateParticipationJob < BaseJob
       def perform(participation_id:, timestamp:)
         call_service!(FranceTravailApi::CreateParticipation, participation_id: participation_id, timestamp: timestamp)
       end
