--- app/jobs/outgoing_webhooks/france_travail/locked_and_ordered_job_base.rb
@@ -1,11 +1,0 @@
-module OutgoingWebhooks
-  module FranceTravail
-    class LockedAndOrderedJobBase < ApplicationJob
-      include LockedAndOrderedJobs
-
-      def self.job_timestamp(timestamp:, **)
-        timestamp
-      end
-    end
-  end
-end
